# openapi.yaml
openapi: 3.1.0
info:
  title: SmeAccounting API
  version: 1.0.0
  description: SME Accounting System (TT99/2025/TT-BTC & Decree 123). Clean Architecture + API-First. Agnostic UI consumes this contract only.
servers:
  - url: /
    description: Local
...
paths:
...
  /api/v1/journal-entries:
    get:
      tags: [JournalEntries]
      summary: List journal entries
      operationId: listJournalEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: fromPostingDate
          in: query
          schema: { type: string, format: date }
        - name: toPostingDate
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paged entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedJournalEntryDto'
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [JournalEntries]
      summary: Create draft journal entry
      operationId: createJournalEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostJournalEntryRequest'
      responses:
        '201':
          description: Created draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseJournalEntryDto'
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Validation failed (RFC 7807)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    Unprocessable:
      description: Domain invariant violated, e.g. GL.UNBALANCED, ACCOUNT.CLOSED_PERIOD (RFC 7807)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    JournalLineDto:
      type: object
      required: [accountCode, debitAmount, creditAmount]
      properties:
        accountCode: { type: string, example: '111' }
        debitAmount: { type: number, format: decimal, example: 1000000 }
        creditAmount: { type: number, format: decimal, example: 0 }
        description: { type: string, nullable: true }
    JournalEntryDto:
      type: object
      required: [id, voucherNo, postingDate, status, lines]
      properties:
        id: { type: string, format: uuid }
        voucherNo: { type: string, example: 'CT-2026-0001' }
        postingDate: { type: string, format: date }
        status: { type: string, enum: [Draft, Posted, Voided] }
        lines:
          type: array
          items: { $ref: '#/components/schemas/JournalLineDto' }
    PostJournalEntryRequest:
      type: object
      required: [voucherNo, postingDate, lines]
      properties:
        voucherNo: { type: string, maxLength: 30 }
        postingDate: { type: string, format: date }
        lines:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/JournalLineDto' }
...
    ApiResponseJournalEntryDto:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/JournalEntryDto' }
...
    PagedJournalEntryDto:
      type: object
      required: [items, page, pageSize, totalCount]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/JournalEntryDto' }
        page: { type: integer }
        pageSize: { type: integer }
        totalCount: { type: integer }
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
        errorCode:
          type: string
          example: GL.UNBALANCED